Optimize memory allocation during source polling
"Approved" by flamegraph results. On my laptop, SystemTime::now() is
surprisingly efficient! It takes very little time, so I don't optimize
it for the moment (we could call it only on flush, and most of time simply do time += poll_interval, at the expense of getting the true timing, but it would be ok for small intervals and fast sources)