plugin-rapl: Recover from almost every situation
I've discovered that, on some Linux installations, /sys/devices/power does not exist (Manjaro Linux with kernel 5.13.19 on an Intel-based laptop). It's not a permission problem, and loading some kernel modules (e.g. intel_rapl_common) did not solve the issue. I haven't found the root cause yet, but in any case, the plugin should be robust to such issues.