fix(alumet): Add the name of the prefix symbols to UnitPrefix::from_str
When receiving a `MeasurementPoint` from a `RelayClientPlugin` by the
`RelayServerPlugin`, the `PrefixedUnit` may have a non empty prefix,
which would be equal to its `unique_name`. But the `from_str` was only
able to convert the `display_name` into the corresponding `UnitPrefix`.
It's now fixed by taking both names into account in the `from_str` match-case.