feat(alumet): Add environment variable interpolation for config files
It is now possible to any linux environment variable in the config
files if the variable does exist and isn't empty. You can obviously
escape the `$` symbol with a backslash to use it.

Some tests has been added to ensure that the regex is working correctly and
that a breaking change will never be able to happen.